+++ dwave/gate/qcdl/operations.py
@@ -60,7 +60,7 @@ from typing import Any, TypeAlias
 
 import numpy as np
 
-from . import implementations
+from .. import implementations
 from .components import QCDLModule
 from .exceptions import QCDLUserError
 from .registers import FixedPointRegister, Register
